requirements.txt
Flask
Flask-SQLAlchemy
Pydantic
psycopg2-binary
flake8